+++ R/QR_export.R
@@ -53,32 +53,32 @@ apply_BQ_style <- function(wb,
             sheet = modalities_sheet,
             rule = '=="Bad"',
             style = bad_style,
-            cols = seq_len(ncol(x$modalities)),
-            rows = 1L + seq_len(nrow(x$modalities))
+            cols = seq_len(ncol(x[["modalities"]])),
+            rows = 1L + seq_len(nrow(x[["modalities"]]))
         )
         openxlsx::conditionalFormatting(
             wb = wb,
             sheet = modalities_sheet,
             rule = '=="Good"',
             style = good_style,
-            cols = seq_len(ncol(x$modalities)),
-            rows = 1L + seq_len(nrow(x$modalities))
+            cols = seq_len(ncol(x[["modalities"]])),
+            rows = 1L + seq_len(nrow(x[["modalities"]]))
         )
         openxlsx::conditionalFormatting(
             wb = wb,
             sheet = modalities_sheet,
             rule = '=="Uncertain"',
             style = uncertain_style,
-            cols = seq_len(ncol(x$modalities)),
-            rows = 1L + seq_len(nrow(x$modalities))
+            cols = seq_len(ncol(x[["modalities"]])),
+            rows = 1L + seq_len(nrow(x[["modalities"]]))
         )
         openxlsx::conditionalFormatting(
             wb = wb,
             sheet = modalities_sheet,
             rule = '=="Severe"',
             style = severe_style,
-            cols = seq_len(ncol(x$modalities)),
-            rows = 1L + seq_len(nrow(x$modalities))
+            cols = seq_len(ncol(x[["modalities"]])),
+            rows = 1L + seq_len(nrow(x[["modalities"]]))
         )
 
         # Appliquer les styles aux bordures
@@ -86,8 +86,8 @@ apply_BQ_style <- function(wb,
             wb = wb,
             sheet = modalities_sheet,
             style = border_style,
-            cols = seq_len(ncol(x$modalities)),
-            rows = 1L + seq_len(nrow(x$modalities)),
+            cols = seq_len(ncol(x[["modalities"]])),
+            rows = 1L + seq_len(nrow(x[["modalities"]])),
             gridExpand = TRUE
         )
 
@@ -97,7 +97,7 @@ apply_BQ_style <- function(wb,
             sheet = modalities_sheet,
             style = rowname_style,
             cols = 1L,
-            rows = 1L + seq_len(nrow(x$modalities)),
+            rows = 1L + seq_len(nrow(x[["modalities"]])),
             gridExpand = TRUE
         )
     }
@@ -109,17 +109,17 @@ apply_BQ_style <- function(wb,
             wb = wb,
             sheet = values_sheet,
             style = border_style,
-            cols = seq_len(ncol(x$values)),
-            rows = 1L + seq_len(nrow(x$values)),
+            cols = seq_len(ncol(x[["values"]])),
+            rows = 1L + seq_len(nrow(x[["values"]])),
             gridExpand = TRUE
         )
 
         # Apply BQ style cell Modalities
-        for (id_col in seq_len(ncol(x$values))) {
-            name_col <- colnames(x$values)[id_col]
-            if (name_col %in% colnames(x$modalities)) {
-                for (id_row in seq_len(ncol(x$values))) {
-                    cell_value <- as.character(x$modalities[id_row, name_col])
+        for (id_col in seq_len(ncol(x[["values"]]))) {
+            name_col <- colnames(x[["values"]])[id_col]
+            if (name_col %in% colnames(x[["modalities"]])) {
+                for (id_row in seq_len(ncol(x[["values"]]))) {
+                    cell_value <- as.character(x[["modalities"]][id_row, name_col])
                     cell_style <- switch(
                         cell_value,
                         Bad = bad_style,
@@ -148,7 +148,7 @@ apply_BQ_style <- function(wb,
             sheet = values_sheet,
             style = rowname_style,
             cols = 1L,
-            rows = 1L + seq_len(nrow(x$values)),
+            rows = 1L + seq_len(nrow(x[["values"]])),
             gridExpand = TRUE
         )
     }
@@ -224,13 +224,13 @@ export_xlsx.QR_matrix <- function(x,
     openxlsx::writeData(
         wb = wb_qr,
         sheet = "Modalities",
-        x = x$modalities,
+        x = x[["modalities"]],
         headerStyle = if (auto_format) header_style else NULL
     )
     openxlsx::writeData(
         wb = wb_qr,
         sheet = "Values",
-        x = x$values,
+        x = x[["values"]],
         headerStyle = if (auto_format) header_style else NULL
     )
 
@@ -381,13 +381,13 @@ export_xlsx.mQR_matrix <- function(
             openxlsx::writeData(
                 wb = wb_modalities,
                 sheet = name,
-                x = qr$modalities,
+                x = qr[["modalities"]],
                 headerStyle = if (auto_format) header_style else NULL
             )
             openxlsx::writeData(
                 wb = wb_values,
                 sheet = name,
-                x = qr$values,
+                x = qr[["values"]],
                 headerStyle = if (auto_format) header_style else NULL
             )
             if (auto_format) {
@@ -435,13 +435,13 @@ export_xlsx.mQR_matrix <- function(
             openxlsx::writeData(
                 wb = wb_mqr,
                 sheet = paste0(name, "_modalities"),
-                x = qr$modalities,
+                x = qr[["modalities"]],
                 headerStyle = if (auto_format) header_style else NULL
             )
             openxlsx::writeData(
                 wb = wb_mqr,
                 sheet = paste0(name, "_values"),
-                x = qr$values,
+                x = qr[["values"]],
                 headerStyle = if (auto_format) header_style else NULL
             )
             if (auto_format) {